Move bounding box code into core.
